requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
Django==4.2.17
django-celery-beat==2.5.0
django-cors-headers==4.6.0
django-encrypted-model-fields==0.6.5
django-extensions==3.2.3
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
kombu==5.4.2
MarkupSafe==3.0.2
mysqlclient==2.2.6
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-magic==0.4.27
six==1.17.0
sqlparse==0.5.1
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
